Input::MouseMoved implemented

+++ src/core/math/Point2.cpp
@@ -1,6 +1,6 @@
 #include "Point2.h"
 
-double Point2::distanceTo(Point2 p){
+float Point2::distanceTo(Point2 p){
 	return sqrt( pow(p.x - x, 2) + pow(p.y - y, 2) );
 }
